+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "example-yagss-site",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "A collection of demos of how to build web sites with yagss",
   "scripts": {
     "serve": "node node_modules/yagss/index.js serve",
@@ -20,16 +20,16 @@
   },
   "homepage": "https://github.com/esonderegger/example-yagss-site#readme",
   "dependencies": {
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "recharts": "^1.5.0",
-    "yagss": "^0.1.6"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "recharts": "^1.8.5",
+    "yagss": "^0.2.0"
   },
   "devDependencies": {
-    "eslint": "^5.13.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.17.0"
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.1.0",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.19.0"
   }
 }
