+++ src/examples/02-inline-slideshow/index.md
@@ -1,0 +1,15 @@
+---
+template: Debug
+title: Inline Sideshows
+images:
+  - filename: jari-hytonen-YCPkW_r_6uA-unsplash
+    caption: Just before 6am
+  - filename: marko-blazevic-zBvVuRJ71vU-unsplash
+    caption: Looking out at the mall at 6:30am
+  - filename: ramiz-dedakovic-bmGy-eqrl1k-unsplash
+    caption: The seal up close
+  - filename: sereja-ris-g3B53PbBfwU-unsplash
+    caption: Making sure everything is where it needs to be
+description: A demo of an interactive component using reactive images.
+---
+TKTK
